# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Puraani Jeans APIs",
        "description": "Puraani Jeans API documentation",
        "contact": {
            "email": "developer@puranijeans.com"
        },
        "license": {
            "name": "Apache 2.0"
        },
        "version": ""
    },
    "servers": [
        {
            "url": "http://puranijeans.test/",
            "description": "dev"
        }
    ],
    "paths": {
...
        "/api/categories/{slug}/products": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get Category Products",
                "operationId": "getCategoryProducts",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "category slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/filter/categories/{slug}/products": {
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get Filtered Category Products",
                "operationId": "getFilteredCategoryProducts",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "category slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Set Order Shipment Details",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "filters": {
                                        "description": "filters",
                                        "properties": {
                                            "records_range": {
                                                "description": "Record Range",
                                                "properties": {
                                                    "show_count": {
                                                        "description": "24",
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "price_range": {
                                                "description": "Price Range",
                                                "properties": {
                                                    "min": {
                                                        "description": "1",
                                                        "type": "string"
                                                    },
                                                    "max": {
                                                        "description": "1",
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "sort_by": {
                                                "description": "Sort by",
                                                "properties": {
                                                    "newest_arrival": {
                                                        "description": "1",
                                                        "type": "integer"
                                                    },
                                                    "featured": {
                                                        "description": "1",
                                                        "type": "integer"
                                                    },
                                                    "price_high_to_low": {
                                                        "description": "1",
                                                        "type": "integer"
                                                    },
                                                    "price_low_to_high": {
                                                        "description": "0",
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "user_access_token": [],
                        "locale": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "user_access_token": {
                "type": "apiKey",
                "description": "user access token",
                "name": "Authorization",
                "in": "header"
            },
            "client_id": {
                "type": "apiKey",
                "description": "client id",
                "name": "x-client-id",
                "in": "header"
            },
            "client_secret": {
                "type": "apiKey",
                "description": "client secret",
                "name": "x-client-token",
                "in": "header"
            }
        }
    }
}